airflow-docker/requirements.txt
pandas
sqlalchemy
requests
flatten-dict
apache-airflow-providers-mysql
apache-airflow-providers-postgres
